requirements.txt
Django==1.11.5
pytz==2017.2
pip==9.0.1
setuptools==28.8.0
wheel==0.30.0
Pillow==4.2.1
django-ckeditor==5.3.0
django-asset==0.1.1
image==1.5.15
olefile==0.44
